fix(NaiveMenuLink): Treat nested paths as parent path for active item
This effects active menu item. A nested non-registered route will be treated as the parent registered route.
For example, a parent route `/books` is registered. Nested routes can be `/books/{book_id}` which are not registered. Now when hitting these routes the menu item of path `/books` will be active.