fix: avoid bg color override of non naive-ui buttons
closes(#56)

Tailwind by default resets the background color to `transparent`. By explicitly specifying this property with `var(--n-color)` we ensure the issue is fixed on `n-button` component. However for `button` element the `var(--n-color)` is resolved from parent element which may lead to unexpected styling.

Note: as this issue's cause depends on the project, from now on it's better to solve it by the user. This is may effect (#8).